# server/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Covid - 19 India Data",
      "version": "1.0.0",
      "title": "Covid-19 Tracker"
    },
    "host": "localhost:3030",
    "basePath": "/api",
...
    "schemes": [
      "http",
      "https"
    ],
    "paths": {
...
      "/getAllCount": {
        "get": {
          "tags": [
            "count"
          ],
          "summary": "Get Active, Recovered, Deceased count",
          "description": "",
          "operationId": "getAllCount",
          "produces": [
            "application/json",
            "application/xml"
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Count"
              }
            },
            "400": {
              "description": "Invalid supplied"
            },
            "404": {
              "description": "Not found"
            },
            "422": {
                "description": "Unprocessable Entity"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Count": {
        "type": "object",
        "properties": {
            "body": {
                "items": {
                    "$ref": "#/definitions/Body"
                  }
            },
            "data": {
                "recoveredCount": {
                    "type": "number"
                  },
                  "activeCount": {
                    "type": "number"
                  },
                  "deceasedCount": {
                    "type": "number"
                  }
            }
        }
      },
...
      "Body": {
        "type": "object",
        "properties": {
                "success": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  }
        }
      }
    }
  }